docker-image-watch: use image digest of the running container
We may have multiple containers running with the same image. If that's
the case, we need to restart all of those containers. In order to do so,
fetch the image digest from the running container instead of checking
the local stamp file.

Also set the default `RESTART_CMD` to restart the container with
systemctl. This will only be run if `AUTOPULL` is set to 1.